+++ libfindchars-examples/src/main/java/zz/customname/FindLiteralsAndPositions.java
@@ -41,7 +41,7 @@ class FindLiteralsAndPositions {
 
             for (int i = 0; i < match.size(); i++) {
                 byte lit = match.getLiteralAt(matchStorage, i);
-                int pos = match.getPositionAt(matchStorage, i);
+                long pos = match.getPositionAt(matchStorage, i);
 
                 if (lit == STAR) {
                     System.out.println("* at: " + pos);
